requirements.txt
# Web Framework
fastapi==0.104.1          # Modern API framework
uvicorn==0.24.0           # ASGI server to run FastAPI
python-multipart==0.0.6   # For file uploads
aiofiles==23.2.1          # Async file I/O for streaming uploads

# Database
psycopg2-binary==2.9.9    # PostgreSQL adapter
sqlalchemy==2.0.23        # ORM (Object-Relational Mapping)
pgvector==0.2.4           # Vector operations in PostgreSQL

# Task Queue
celery==5.3.4             # Background task processing
redis==5.0.1              # Redis client
flower==2.0.1             # Celery monitoring UI

# AI/ML
openai==1.3.7             # OpenAI API client
tiktoken==0.5.2           # Token counting for chunking

# File Processing
pypdf2==3.0.1             # PDF text extraction
python-docx==1.1.0        # DOCX text extraction

# Utilities
pydantic==2.5.0           # Data validation
pydantic-settings==2.1.0  # Settings management
python-dotenv==1.0.0      # Load .env files

# ML/AI
numpy==1.24.3             # Numerical computing
scikit-learn==1.3.0       # Machine learning utilities

# Testing
pytest==7.4.3             # Testing framework
httpx==0.25.2             # Async HTTP client for tests

# Local AI Models (NEW)
sentence-transformers==2.2.2  # Local embeddings
transformers==4.35.2
huggingface-hub==0.19.4
torch==2.1.1                  # PyTorch for local models